Hacks the davinci-mcasp to be slave and master
The alsa-soc architecture doesn't tailor to having the receive section
be slave, and the transmit section be master. Furthermore, both sections
needed to be clocked from an external source, in my case by a recovered
clock from a DIR9001 optical receiver.

This make sure the output is always exactly in sync with the externally
clocked input signal, and allows playback of it sample-for-sample
without getting out of sync. It also has the added benefit of supporting
any sample rate (that is up to 48 kHz, because I chose the 512 * frame clock to
support 8 channel TDM output).

With this hack, the driver ignores any sample rate option you give, both
for capture and playback. Clock dividers are hardcoded, because how I
use it, the master clock is decoded from a bi-phase signal by the
DIR9001, and is always 512 times the frame clock.

I also hardcoded a receive TDM slot number of 2. The driver was only
tailored for synchronous receive/transmit, but the hardware doesn't have
this restriction.